python_env_requirements_docker.txt
torch-cluster
torch-geometric
torch-scatter
torch-sparse
torch-spline-conv
torchaudio

bitstring
nni